+++ src/form.js
@@ -1,5 +1,6 @@
 'use strict';
 var browserCookies = require('browser-cookies');
+var formUtility = require('./form-utils');
 var MYDATE = 9;
 var MYMONTH = 6;
 var DEFAULTMARK = 3;
@@ -41,26 +42,12 @@ var DEFAULTMARK = 3;
 
   submitButton.onclick = function(evt) {
     evt.preventDefault();
-    var expires = getExpires();
+    var expires = formUtility.getExpires(MYDATE, MYMONTH);
     browserCookies.set('mark', selectedReviewMark, {expires: expires});
     browserCookies.set('name', reviewName.value, {expires: expires});
+    console.log(expires);
   };
 
-  function getExpires() {
-    var now = new Date();
-    var currentDate = now.getDate();
-    var currentMonth = now.getMonth();
-    var currentYear = now.getFullYear();
-    var myYear;
-    if (currentMonth >= MYMONTH && currentDate >= MYDATE) {
-      myYear = currentYear;
-    } else {
-      myYear = currentYear - 1;
-    }
-    var birthday = new Date(myYear, MYMONTH, MYDATE);
-    var difference = Math.floor((now - birthday) / (1000 * 60 * 60 * 24));
-    return difference;
-  }
 
   for (var y = 0; y < reviewMark.length; y++) {
     reviewMark[y].onclick = function() {
