+++ src/main.js
@@ -1,4 +1,5 @@
 'use strict';
+
 require('./reviews');
 require('./form');
 require('./game');
